+++ src/wm/windowlistbox.cpp
@@ -226,10 +226,10 @@ int cWindowListBox::GetInt(int index)
 
 void cWindowListBox::click()
 {
-	int xx = mousex;
+	int xx = (int)mousex;
 	xx -= realx();
 	xx -= parent->px();
-	int yy = Graphics.h()-mousey;
+	int yy = Graphics.h()-(int)mousey;
 	yy -= realy();
 	yy -= parent->py();
 
@@ -281,10 +281,10 @@ void cWindowListBox::click()
 
 void cWindowListBox::drag()
 {
-	int xx = mousex;
+	int xx = (int)mousex;
 	xx -= realx();
 	xx -= parent->px();
-	int yy = Graphics.h()-mousey;
+	int yy = Graphics.h()-(int)mousey;
 	yy -= realy();
 	yy -= parent->py();
 
@@ -303,7 +303,7 @@ void cWindowListBox::drag()
 		if (barpos != 0)
 			barpos = (int)((float)(h-16.0f) * ((float)liststart / (float)values.size()));
 
-		int offset = yy-dragoffsety;
+		int offset = (int)(yy-dragoffsety);
 
 		barpos-=offset;
 		if (barpos < 0)
